Accept only valid menu options in boot menu
Booting from USB did not work when no FTDI cable was connected. After adding
a pull-up on RXD it did work. This lead to the idea that appearantly the input
is considered low when not connected and the pull up fixed that. If the input
is low a character is received, even though it being a break. The menu
code considered it an "invalid choice" and starts to boot the Linux kernel
(default option). This change makes the menu code ignore any invalid character
received on the UART. Only keys in the menu are considered valid.